made userid an int, which ended up touching a lot
Former-commit-id: 3955fd43fdba629a9688a2c011c344a6e5a21bb7 [formerly a458a8360b7c48e57b6d594f798c50f0951f6aa8] [formerly 335c86dc7e2cb36ca6badbdefdefa5fab7547602 [formerly 60ca7913af5e236a6d015e3bd761629d75930d73]]
Former-commit-id: d15f4f4c3e556947f3e5afaa528b03c870ce862a [formerly 9a6c609a0fbd9287b830591b0c4cce202c36691c]
Former-commit-id: c3de9414b6723845da6c9d7d2dec4f5d3fc30cd0

+++ src/mitll/langtest/server/DownloadServlet.java
@@ -216,7 +216,7 @@ public class DownloadServlet extends DatabaseServlet {
     if (!englishPart.isEmpty()) englishPart = "_" + englishPart;
 
     // user part
-    String userPart = getUserPart(db, Long.parseLong(useridString));
+    String userPart = getUserPart(db, Integer.parseInt(useridString));
 
     String fileName = foreignPart + englishPart + userPart;
     fileName = fileName.replaceAll("\\.","");
@@ -228,7 +228,7 @@ public class DownloadServlet extends DatabaseServlet {
     return underscores;
   }
 
-  private String getUserPart(DatabaseImpl db, long userid) {
+  private String getUserPart(DatabaseImpl db, int userid) {
     User userWhere = db.getUserDAO().getUserWhere(userid);
     return userWhere != null ? (userWhere.getUserID().isEmpty() ? "":"_by_" + userWhere.getUserID()) : "";
   }
